+++ scvi/core/lightning/__init__.py
@@ -1,4 +1,4 @@
-from .vaetask import VAETask, SemiSupervisedTask, AdvesarialTask
-from .trainer import Trainer
+from ._vaetask import VAETask, SemiSupervisedTask, AdvesarialTask
+from ._trainer import Trainer
 
 __all__ = ["VAETask", "Trainer", "SemiSupervisedTask", "AdvesarialTask"]
