+++ scvi/core/lightning/_progress.py
