+++ scvi/core/lightning/_vaetask.py
@@ -10,7 +10,7 @@ from torch.optim.lr_scheduler import ReduceLROnPlateau
 from scvi._compat import Literal
 from scvi.core.modules import Classifier
 from scvi.core.modules._base._base_module import AbstractVAE
-from scvi.core.modules.utils import one_hot
+from scvi.core.modules._utils import one_hot
 from scvi import _CONSTANTS
