+++ scvi/core/models/_base_model.py
