+++ scvi/core/models/_rnamixin.py
