+++ scvi/core/models/_vaemixin.py
