+++ scvi/core/modules/__init__.py
@@ -1,8 +1,8 @@
-from .autozivae import AutoZIVAE
-from .classifier import Classifier
-from .jvae import JVAE
-from .scanvae import SCANVAE
-from .totalvae import TOTALVAE
-from .vae import LDVAE, VAE
+from ._autozivae import AutoZIVAE
+from ._classifier import Classifier
+from ._jvae import JVAE
+from ._scanvae import SCANVAE
+from ._totalvae import TOTALVAE
+from ._vae import LDVAE, VAE
 
 __all__ = ["VAE", "LDVAE", "TOTALVAE", "AutoZIVAE", "SCANVAE", "Classifier", "JVAE"]
