+++ scvi/core/modules/_classifier.py
