--- scvi/core/trainers/__init__.py
@@ -1,9 +1,0 @@
-from .inference import UnsupervisedTrainer
-from .jvae_trainer import JVAETrainer
-from .total_inference import TotalTrainer
-
-__all__ = [
-    "UnsupervisedTrainer",
-    "TotalTrainer",
-    "JVAETrainer",
-]
