+++ scvi/core/utils/__init__.py
@@ -1,3 +1,3 @@
-from .differential import DifferentialComputation
+from ._differential import DifferentialComputation
 
 __all__ = ["DifferentialComputation"]
