+++ scvi/core/utils/_differential.py
