+++ scvi/model/_linear_scvi.py
