+++ scvi/model/_scvi.py
