+++ scvi/model/_totalvi.py
